requirements.txt
fastmcp
google-genai
python-dotenv
biopython>=1.80
pytest>=7.0
setuptools
mapchiral
rdkit
cobra
git+https://github.com/JBEI/RetroTide.git
requests
beautifulsoup4
streamlit